Remove unnecessary noop handlers.
